`Publisher#flatMapSingle()` may complete prematurely
__Motivation__

Successful completion of constituent `Single`s and main `Publisher` may trigger a race in which completion event is enqueued before success from the `Single`.
This happens because we decrement the active `Single`s counter before enqueuing the result of the `Single` to the data queue.
When the outer `Publisher` completes, we check if the active count is 0 then we emit the completion event to the queue.
These two events race and hence we may miss the results of the `Single` that have not yet been enqueued to the queue but the count have been decremented.

__Modification__

Now decrement counter after enqueuing data in the queue.

__Result__

No more missed results.