Ignore cancel on transacted command futures.
Motivation:

Cancelling the `Future` returned by Redis transacted command methods may
result in an inconsistency between the commands sent to the wire, and the
`SingleProcessor`s added to the list. To avoid this, we should ignore
cancels on the returned `Future`.

Modifications:

Ignore the cancel for command methods.

Results:

Command methods can no longer be cancelled, avoiding potential
inconsistencies.