Clarify NettyIoExecutors' 0 thread number behaviour (#142)
## Motivation

We use `0` as the number of thread passed to the internal `NettyIoExecutors` but the the behaviour for `0` is not documented. We also use the internal `NettyIoExecutors` in a places where we should use the public one, where a method overload allows not providing a thread count (instead of using `0`).

## Modifications

- Add overload without thread number in internal `NettyIoExecutors` and use it in public `NettyIoExecutors`.
- Use the public `NettyIoExecutors` where it makes sense.

## Results

Less magic numbers.