`DefaultDnsServiceDiscoverer` prematurely decrements requested count (#154)
__Motivation__

`DefaultDnsServiceDiscoverer` will eventually stop querying DNS if there is no change in the DNS entries.
It decrements `pendingRequests` count even if it did not emit any event (no change in DNS entries). This will make `pendingRequests` `0` after some polls.
Downstream `Subscriber` MAY not request more if no items were emitted. This means we will stop polling if the requested amount is finite.

__Modification__

Decrement `pendingRequests` only when we are emitting an event.

__Result__

Correct demand management and keep polling as long as there is demand.