Correctly cascade result when extending `CompletableFuture`
__Motivation__

If a `CompletableFuture` is returned from any of the functions on a `CompletableFuture` (eg: `thenApply()`) then the JDK implementation of `CompletableFuture` does not add a listener to it.
Our implementations of `CompletableFuture` do not add a listener to the wrapped `CompletableFuture` and hence does not get completed if the wrapped future completes.
So, when used by another JDK `CompletableFuture`, we may never populate the fields in `CompletableFuture` which indicates termination.

A simple code example is:

```java
CompletableFuture.completedFuture("foo")
        .thenCompose(s -> Single.success(s + "bar").toCompletionStage())
        .get()
```

__Modification__

When wrapping another `CompletionStage`, add a listener to populate the result to the returned `CompletableFuture`.
For methods that populate the result, also populate self result (not just for the delegate)

__Result__

State of the `CompletableFuture` is correctly updated.