Improper handling when `subscribe()`` throws (#159)
__Motivation__

The specification says that `Publisher.subscribe()` should not throw but for cases when `Subscriber` is `null`.
We do work before we delegate the call to actual asynchronous source implementation. We should protect against errors coming out of that work, eg: if we are offloading, an `Executor` may throw.

__Modification__

If there is an error inside `subscribe` we assume that the `Subscriber` has not been invoked and we send the error to the `Subscriber`.

__Result__

Correct `Subscriber` termination even when `subscribe()` throws.