Improve behavior for `DoBeforeSubscribe` in presence of exceptions
__Motivation__

`do[Before/After]Subscribe()` behavior is not ideal for the following scenarios:

- `doBefore` tries to send the `onSubscribe()` even if the callback failed. It also throws from `onSubscribe`. Since we are throwing from `onSubscribe()` we expect the source to bail on the `Subscriber` so we should reduce the exposure of the `Subscription`/`Cancellable.
- `doAfter` call the callback even if the original `Subscriber` threw from `onSubscribe()`. We do not need to do anything here since the original `Subscriber` has violated specifications.

__Modification__

Do not do any exception handling in `do[Before/After]Subscribe()` and let the original source decide how to handle these exceptions.

Also removed usage of `$` as the variable name in callback when they are ignored.

__Result__

Reduced complexity and less expsoure of `Subscription`/`Cancellable` in presence of spec violations.