Publisher#flatMapSingle code cleaning
Motivation:
Publisher#flatMapSingle was not propagating signals in some instances when cancel was invoked. This is not required, and can obfuscate control flow in some instances. We should instead let the upstream Publisher determine the behavior of cancel instead of trying to filter signals in operators.

Modifications:
- Remove code associated with cancel preventing the delivery of signals.

Result:
Less code and assumptions about cancel in Publisher#flatMapSingle.