Add `collect()` for `Single` and `Completable`
__Motivation__

Today, in order to asynchronously collect a known set of `Single`/`Completable`, one has to use `Publisher`.
This creates friction as many users are not comfortable dealing with streaming APIs. We can add a shorthand for `collect()` to help such users.

__Modification__

Added static `collect()` variants for `Single` and `Completable`
Also added `Publisher#flatMapCompletable()` as that is required for `Completable#collect()`

__Result__

Less friction for scalar API users.