Correctly assert closed-connection exception
__Motivation__

`InsufficientlySizedExecutorHttpTest` asserts whether connection is closed due to task rejection.
Depending on which state the channel is when the connection close is detected, netty sometimes throws an `IoException` instead of `ClosedChannelException`.
We should check for either of these exceptions.

__Modification__

Enhance the assertion to optionally look for `IoException`.

__Result__

Better assertions, less flake!