Tighten filters contract
__Motivation__

Today clients and servers allow users to provide any implementation returned from the filters.
This is a loose contract as users may forget to correctly call delegate methods like `closeAsync()`.
This also makes it hard for ServiceTalk to call a method that can traverse through all the filters in the chain as the chain can arbitrarily be broken by one filter that does not correctly delegate.
We can tighten this contract by defining a filter class (same as the adapters we provide today) which is always returned from these filter factories.

__Modification__

- Provide adapters for all filters.
- Have the filter factories return the filter instance defined above.
- For HTTP only provide filter contract for streaming and not for other programming model as we expect people to only provide streaming filters and use the converters to convert to/from the streaming filters.

__Result__

Tighter filter contract and hence capability for ST to inspect all filters in the chain.