Single to CompletionStage propagate status (#207)
Motivation:
SingleToCompletionStage converts from Single to CompletionStage. However the CompletionStage listener methods may have only listened for one of the respective outcomes (e.g. success or failure) in order to invoke the respective callback methods. The listeners always need to handle both outcomes (e.g. success and failure) and propagate the outcome to the next stage to ensure a continue chain is maintained. Otherwise this may lead to hung stages which never complete if the unhandled status is not propagated.

Modifications:
- SingleToCompletionStage internal listener must always handle success and failure, and always propagate the status.
- SingleToCompletionStage intennal listener generics are made more strict to get better type safety.

Result:
No more hung CompletionStage chains if the status is no propagated to the CompletionStage listener used.