Redis encoder improvements (#217)
Motivation:

Writing CharSequences using Netty's `ByteBufUtil.reserveAndWriteUtf8` is faster. However, this means we need to know the exact number of bytes before we write (since we no longer have the byte array). As such, we use
`StringByteSizeUtils.numberOfBytesUtf8` to correctly calculate the number of bytes. Writing all CharSequences this way is more performant, so we can
replace estimating the length with calculating it correctly.

Modifications:

Call `calculateRequestArgumentSize` instead of `estimateRequestArgumentSize`
for `CharSequences`, other than those in a `Collection`.

Results:

Improved performance.