Fix string byte length calculation (#235)
Motivation:

`numberOfBytesUtf8` is used to specify, in the Redis protocol, how many
bytes will be written for a bulk string, so it must be correct. It was
handling surrogates (4-byte characters) incorrectly.

Modifications:

Stop over-counting by 1 for 4-byte characters.

Results:

String byte lengths are correct for strings with 4-byte characters.