Refactor HTTP and Redis retry filters
Motivation:

1. In retry-filters we check the request meta-data/command up front to
understand if the request is retryable or not. Instead, we should have
the request meta-data and throwable available for `retryWhen` operator.
2. `RetryingHttpRequesterFilter` does not apply `retryWhen` operator for
a reserved HTTP connection, `RetryingRedisClient` does not applicable
for `RedisConnection`s at all.
3. Builders are a bit buggy and does not provide a way to compose a
retry strategy to consider request meta-data and throwable at the same
time or compose it with users logic on top.
4. `RetryingHttpRequesterFilter` and `RetryingRedisClient` are inconsistent.

Modifications:

- Always apply `retryWhen` operator in `request` method, including a
reserved connection;
- Use `AbstractRetryingFilterBuilder` to avoid code duplication;
- Accept a `BiPredicate` for `HttpRequestMetaData`/`Command` and
`Throwable` as a retry-for logic;
- Rename some builder methods and parameters to align with
`RetryStrategies` terminology;
- Set the default value for `maxRetries` to 3;
- Enable exponential backoff with jitter by default;
- Make the default retry-for predicate public to let users wrap/enhance it;
- Add `ReservedRedisConnectionFilter`;
- Remove unused `RedisRequestAwareRetryStrategy`;
- Add missed `toString()` overrides;
- Improve javadoc;
- Fix tests;

Result:

Correctly applies `retryWhen` operator and provides more flexible and
consistent configuration for `Retrying(Http/Redis)RequesterFilter`.