Fixing a race condition in a DNS service discovery test. (#262)
Motivation:

The A record for `apple.com` was inadvertently removed from a test. This
resulted in an `onError` for resolution, while the test was depending on
`onError` due to closure. This lead to a race.

Modifications:

Add back an A record for `apple.com` in the test.

Results:

The test passes consistently.