`noOffloadsStrategy()` should not return `immediate` `Executor`
__Motivation__

Currently `ExecutionStrategy` returned by `noOffloadsStrategy()` returns an `Executors.immediate()` from its `executor()` method.
This when used with an Http service may set the server `Executor` (as contained by `ExecutionContext`) as `immediate()`.
This may be surprising if this `Executor` is used for creating any asynchronous resource as the user's intention is to turn off offloading done by ST.

__Modification__

 - `noOffloadsStrategy()` returns `null` from it's `executor()` method.
 - Added a strategy adapter class for HTTP and redis to easily override the behavior of `noOffloadsStrategy()` to actuallly return a different `Executor`.

 __Result__

 More predicatable behavior when using `noOffloadsStrategy()`.