Add unit test for HttpResponseDecoder for empty line trailer terminator
Motivation:
There was a bug recently found in Netty where the decoder would discard a
trailer if there was an empty line that terminates the trailers was parsed in a
separate buffer.

Modifications:
- Add a unit test to HttpResponseDecoderTest to verify this scenario is handled
correctly.

Result:
Better test coverage for HttpResponseDecoder, and confirmation we are not
subject to the same bug.