MaxRequestLimitExceededException exception from closed connection is confusing
__Motivation__

With an HttpConnection (not HttpClient), if a request is sent on a closed connection, instead of a ClosedChannelException (or any IOException), users receive MaxRequestLimitExceededException. This is somewhat confusing because it’s not an IOException, it’s a RetryableException.
By combining connection status and transportError we can provide more clear error messages to the user.

__Modifications__

- enhance `RequestConcurrencyController` to allow exposing an unavailable connection state
- enhance `ConcurrentRequestsHttpConnectionFilter` to leverage transport errors and connection state

__Result__

Users won't get confusing error messages suggesting wrong configuration when the connection is closed.