Reduce `Await` usage.
__Motivation__

`Await` was created when we did not have `Future` conversion for our sources.
Since we added those conversions, we can remove usages of `Await` in favor of `toFuture().get()`.

__Modification__

For test code we use some methods of `Await` which are useful, eg: `awaitNonNull` and `awaitIndefinitelyUnchecked`.
This change does the following:
- Remove usage of `Await` in production code.
- Remove all the custom implementation inside `Await` and use `toFuture().get()` instead.
- Move `Await` to `testFixtures`
- Rename `Await` to `BlockingTestUtils`

__Result__

Reduce usage of `Await` and remove custom code.