SpliceFlatStreamToMetaSingle should warn when errors may not be delivered
__Motivation__

Due to the nature of this operator it's possible that errors observed from the source after MetaData was delivery (successfully or not) won't ever be delivered if nobody subscribes to the Publisher inside the MetaData. This would happen for example if delivery of the Single<MetaData> throws an error.

__Modifications__

Log a debug message when an error is queued up without anyone listening.

__Result__

Better visibility and debuggability for users of this operator