Conditional filters decouples async context
__Motivation__

Conditional filters create a `Single` but do not share the context on subscribe.
This makes it such that the context within a filter chain is decoupled from the previous filters and hence changes made to the context after a conditional filter will not be available to the filters before it.

__Modification__

We decided initially to make these filters such that the predicates are invoked on each subscribe.
The benefit was considered to be simplicity of predicate that it is invoked each time a request is sent.
Considering the complexity of managing contexts and an additional layer of indirection, it seems we can avoid this as we expect predicates to be evaluated based on immutable request properties and hence the result of the predicate must be idempotent.

__Result__

Conditional filters now preserve the context mutations.