Fixing repeat test (#330)
Motivation:

This test has been passing incorrectly.
Because the `repeat` was not repeating, the `onComplete` propagated to the
subscriber. `verifySubscribed` is passing only because the `TestPublisher`
is created with `preserveSubscriber`.
The intent of this test (based on the name) is to ensure that the 3rd item,
requested before the `onComplete,` is still requested after the repeat.

Modifications:

- Enable repeat for the test
- Assert that the subscriber does not receive any termination signals.

Results:

Test passes for the right reason.