ConnectionAcceptor simplification
Motivation:
The ConnectionAcceptor requires an additional concrete class ConnectionAcceptorFilter and factory (e.g. ConnectionAcceptorFilterFactor) in order to use. Following this pattern is nice because it follows a familiar pattern as the client/service filters, but in this case it creates overhead to use and understand. The additional functionality provided by the Filter and Factory is the ability to append logic together, but we can add this directly to the ConnectionAcceptor interface.

Modifications:
- Delete the ConnectionAcceptorFilter and ConnectionAcceptorFilterFactor APIs
- Add the ConnectionAcceptor#appen(ConnectionAcceptor) API that allows for logic composition

Result:
Simplification of the ConnectionAcceptor API.