Adding script to help keep copyright dates updated. (#371)
Motivation:

When more than trivial changes are made to a file, the year(s) in the
copyright header should be updated with the current rule. Automating the
process of updating, with user-input for deciding what qualifies as
"more than trivial", makes this process easier.

Modification:

Add script for checking and optionally updating copyright dates on
modified files.

Results:

Keeping copyright dates updated is easier.