TestExecutor scheduled time management
Motivation:
TestExecutor initializes its time to a random number to simulate the behavior of
System.nanoTime(). A ConcurrentNavigableMap is used to store all scheduled tasks
and an offset from the currentNanos is taken in an attempt to get the next set
of Runnables that have expired and should be executed. However if the initial
value is sufficiently high such that the scheduled task offset would overflow
this approach will not work because the ConcurrentNavigableMap lookup based upon
currentNanos won't give the correct next set of tasks to execute.

Modifications:
- We should transpose the epoch time onto Long.MIN_VALUE so that overflow only
  occurs when the full domain of the long type is exhausted. This requires that
when ever we use time for managing scheduled tasks we use the transposed time
instead of the raw currentNanos.
- executeNextScheduledTask should get the first entry, which should always be
  the next to expire and ready to execute. If a task is executed, we should
check if the queue is empty and remove it from the underlying
ConcurrentNavigableMap.

Result:
More correct time management in TestExecutor for scheduled tasks.