Improving async operator generics.
Motivation:

We have operators methods that have generic arguments, however we were not
using the appropriate wild cards on the generics. The impact of this would
be if users have a Function (or some generic type) they may have to convert
the types manually when the method signature could otherwise support their
type.

Modifications:

Add `? extends` to generics in operator method signatures and
implementations as appropriate.

Results:

Operator API allows for subclasses without having to convert types
manually.