Compare HttpResponseStatus, HttpProtocolVersion, HttpRequestMethod properly (#370)
Motivation:

We recently removed enums as implementation for `HttpResponseStatus`,
`HttpProtocolVersion`, and `HttpRequestMethod` and refactored these 
interfaces to convert them to final classes. We should revisit the way we
compare instances of these classes, because sometimes we use `==` and
`!=` instead of `equals` method.

Modifications:

- Use `equals` to compare `HttpResponseStatus`,`HttpProtocolVersion`,
and `HttpRequestMethod` instances;
- Use `HttpResponseStatus.code()` and `HttpRequestMethod.name()`
for comparison in internal code as optimization;
- Use reference comparison for `HttpRequestMethod` in HTTP decoders;

Result:

`HttpResponseStatus`, `HttpProtocolVersion`, `HttpRequestMethod`
correctly use `equals` instead of `==` and `!=`.