Conditional[Client|Service|Connection] filter lifetime management (#403)
Motivation:
The Conditional[Client|Service|Connection] filter implementations were not managing the lifetime of the AsyncClosable entity that is invoked when the predicate succeeds. The entity is generated from a filter factory create() call on the client side, or provided by the user on the server side, and either way should be owned by the Conditional filter.

Modifications:
- Conditional[Client|Service|Connection] filter to take ownership of the predicate entity

Result:
Conditional[Client|Service|Connection] filter don't leak the lifetime of the predicate entity.