Enhance LoggerStringWriter to wait for log content to stabilize
Motivation:
Some tests may log from multiple threads and want to assert the content that is ultimatley logged. LoggerStringWriter doesn't provide any utilities to wait for logging content to stabilize before using the accumulated log data. This means we may try to validate partially complete logs and fail tests as a result.

Modifications:
- Add a stableAccumulated that will wait for the current thread logs to be flushed, and also attempt to wait for other threads too.

Result:
LoggerStringWriter provides the ability to wait for logging content to be flushed.