Remove custom CompletionStage implementation (#398)
Motivation:
ExecutorCompletionStage was added to preserve the Executor for all listeners of
the Future chain. However this is generally not expected by users of the
CompletionStage API and not required by the API. This approach may actually be
counterproductive if offloading is enabled as unnecessary offloading may occur.
For example if someone does a Future#get() after the future has already
completed we will offload to notify this listener, even though the user's code
will always execute on the calling thread.

Modifications:
- Remove ExecutorCompletionStage and related CompletionStage classes
- Only retain a CompletableFuture implementation and wrappers for AsyncContext.
  This also has the advantage that if folks need CompletableFuture we don't have
to force additional allocations.

Result:
Fixes https://github.com/servicetalk/servicetalk/issues/348.