CompletableProcessor and SingleProcessor may retain Subscriber after cancel
Motivation:
CompletableProcessor and SingleProcessor call onSubscribe before adding the
Subscriber to the listener queue to ensure we respect the Subscriber contract.
However if the Subscriber cancels synchronously we will still add the Subscriber
to the listener queue, and won't deference the Subscriber until the source
completes (which may take very long or never happen). This means the Subscriber
will not be eligible for GC when it should be.

Modifications:
- CompletableProcessor and SingleProcessor now use a DelayedCancellable so that
we can call onSubscribe up front, but remove from the listener queue if cancel
happens synchronously.

Result:
No more memory leak in the event of synchronous cancel from CompletableProcessor
and SingleProcessor.