`Completable` when converted to a value containing source (`Single`, `Future`, etc) should use `Void` (#436)
__Motivation__

Currently `Completable` methods to convert to `Future`, `CompletionStage` and `Single` return those types with the generic parameter as `T`.
The methods then emit `null` as the value. This is potentially confusing as we are forcing `null` and may introduce bugs when someone inadvertently uses the `null` value.

__Modification__

Converted these methods to now use `Void` as the generic parameter. `Void` is true to what we are emitting; `null`

__Result__

More intutitive and safer conversions.