Decouple StreamingHttpService from StreamingHttpServiceFilter
Motivation:
The introduction of HttpExecutionStrategy and the executionStrategy() on the
StreamingHttpService API has created some ambiguity in the service and filter
APIs. The intention is that the StreamingHttpService can provide an
HttpExecutionStrategy and this will be "merged" with the filter chain to compute
the effective strategy. Through this process the StreamingHttpServiceFilter API
should not directly be providing an HttpExecutionStrategy via a
executionStrategy() method, but instead they should have an opportunity to merge
the HttpExecutionStrategy introduced by that filter (if any) and propagate that
to other filters so that an overall "effective strategy" can be computed
considering all filters and the service.

Modifications:
- StreamingHttpService (and other HttpService APIs) are now interfaces which
  retain the executionStrategy() method.
- StreamingHttpRequestHandler is now AsyncClosable
- StreamingHttpServiceFilter extends from StreamingHttpRequestHandler and
  delegates to a StreamingHttpRequestHandler
- HttpServiceFilterFactory takes a StreamingHttpRequestHandler and returns a
  StreamingHttpServiceFilter. This allows for avoiding unnecessary wrapping just
  to accommodate the "effective" HttpExecutionStrategy strategy. This
computation can be done a single time at build time and can fail fast if
incompatibilities are detected.

Result:
StreamingHttpService and StreamingHttpServiceFilter are decoupled and their APIs
are refined for their purpose.